moved component into hook
